Tests refactoring: moved to separate files
