+++ lib/index.js
@@ -7,7 +7,7 @@ var buildSync = require('./build-sync');
 
 /**
  * @param {String} root
- * @returns {{entity: *, is: *, fromJson: *}}
+ * @returns {{entity: *, is: *, fromJson: *, buildSync: *}}
  */
 function instance(root) {
